Remove dependency on before_commit for CI checks (#144)
[before_commit](https://github.com/EnvironmentAgency/before_commit) was an attempt to have a central gem which could be included in every Ruby based project. The gem would manage the checks we wished to perform in each, thereby ensuring consistency and quality across the projects.

In practise we have found when it works, it works great however when it doesn't, it can be very confusing to try and resolve what is actually failing (the project, the check, or **before_commit** itself). We also frequently encountered issues with trying to essentially enforce a centrally controlled workflow on differing development environments.

We have agreed instead to move to ensuring the checks are run independently, and only on the relevant CI server. It will be the responsibility of devs to run these same checks locally, in the way appropriate to their environment, to ensure they don't '**break the build**'.

This change covers removing the dependency on **before_commit** and instead moving to the new agreed model. This involves

- removing **before_commit** references
- removing configuration specific to **before_commit**
- removing usage of **before_commit** from the `.travis.yml` file
- adding back each of the checks **before_commit** was handling, via gems, config files, updates to `.travis.yml` or combinations of them
- adding default task to Rakefile to run both **rubocop** and **spec**
- adding our standard `.rubocop.yml` to the project
- updating `spec_helper.rb` to not use 'doc' format when only running one test (as there is only one test!)
- adding `.rspec` file (was missing from the project)
- removing matchers in `.gitignore` which would prevent files like `.rubocop.yml` from being committed
- resolving any **Rubocop** issues found
- resolving any **i18n-tasks** 'unused' errors found
- remove `.scss-lint.yml` as we are no longer running this linter
- updating documentation (README.md, .env.example) to reflect the changes made

Additionally, though the version of Rubocop used on the project is not currently erroring on this issue, we have taken to opportunity to resolve it.