Bundle Update on 2017-12-15 (#179)
* Bundle Update on 2017-12-15

* Resolve issues with rubocop breaking build

[Deppbot](https://www.deppbot.com/) has updated rubocop and this has caused a break in the build due to new errors being found.

The one cop we have had to disable is `Style/FormatStringToken` which was introduced in 0.49.0.

<https://github.com/bbatsov/rubocop/issues/3438>

The only issue its highlighting in the code is `strftime()`, and we feel the way we are using it is as expected so are choosing to disable this cop for now.

* Update FactoryGirl to FactoryBot

The people behind ThoughtBot have decided to change the name of this gem, which breaks build if we don't rename it in our code. This change updates the gem and the code to use the new name.

* Fix error following rubocop change

We didn't implement a rubocop fix properly. This change corrects that mistake.

* Update flood_risk_engine to latest