Update Travis config (#206)
The build was failing due to our gem installation flags. This new config is closer to what we use on WEX and WCR. It scraps Brakeman as Hakiri already handles these checks for us. We also no longer use overcommit or html-hint so can scrap those.

This PR also updates a few outdated gems which were causing CI issues.