Re-add whenever gem to fix deployments (#222)
This app does not use whenever for scheduling (although the back office does). However our deployment scripts are built around both apps having the gem installed. So until we are in a position to update our deployment process, we need to reinstall the gem or deployments will be broken.

This undoes the work of a few PRs (#152, #221).