Alter airbrake config to allow use in development
The previous team took the decision to not activate airbrake in non-production environments, no doubt to reduce noise.

We on the other hand value having it enabled in all environments as it might highlight issues we haven't spotted. It also reduces the differences between the app running in production, and running in other environments.

So for these reasons we're making these changes, which bring the config more inline with our other services, and ensure Airbrake is always enabled by default irrespective of environment.