Hopefully fix travis build
When we made these same changes to the back office in <https://github.com/DEFRA/flood-risk-back-office/pull/170> the build succeeds. In this project it broke. The issue was an error caused when trying to run the db:migrate rake task but not have an airbrake project key set as an ENV VAR.

A comparison of the `secrets.yml` and the `.travis.yml` between the projects showed neither specified a default, nor is there anything set directly in Travis.

The one difference found is that the back-office's `.env.test` file specifies the required airbrake env vars. The front office version didn't.

So this change is premised on the fact that when rails is run under `RAILS_ENV=test` dotenv automatically reads your `.env.test` file because this matches the environment specified.