Alter airbrake config to allow use in development (#270)
The previous team took the decision to not activate airbrake in non-production environments, no doubt to reduce noise.

We on the other hand value having it enabled in all environments as it might highlight issues we haven't spotted. It also reduces the differences between the app running in production, and running in other environments.

So for these reasons we're making these changes, which bring the config more inline with our other services, and ensure Airbrake is always enabled by default irrespective of environment.

## Remove some of the black-list keys

In our recent work, we do not believe we need to go to this level to strip values from the errors we send to Errbit.

So this change removes some, which are either transient i.e. they are just related to temporary data, or which are likely to appear on the public register for flood risk exemptions.

## Allow Travis to build

When we made these same changes to the back office in <https://github.com/DEFRA/flood-risk-back-office/pull/170> the build succeeded. In this project, it broke. The issue was an error caused when trying to run the `db:migrate rake` task but not having an airbrake project key set as an ENV VAR.

A comparison of the `secrets.yml` and the `.travis.yml` between the projects showed neither specified a default nor is there anything set directly in Travis.

The one difference found is that the back-office's `.env.test` file specifies the required airbrake env vars. The front office version didn't.

So this change is premised on the fact that when rails is run under `RAILS_ENV=test` dotenv automatically reads your `.env.test` file because this matches the environment specified.