Fix airbrake project env var reference (#272)
https://eaflood.atlassian.net/browse/RUBY-604

At the start of 2019 we migrated the FRAE service to a new environment which had a different struture. Previously the front and back office apps were on separate instances. In the new structure they were put on the same one.

Whilst working on the RUBY-604 we've spotted that both apps refer to just one environment variable; `AIRBRAKE_PROJECT_KEY`. This was fine when they were on separate instances, but now there on one it means all exceptions are being logged to the same project in Errbit.

We like to keep a separate Errbit project for each app as it makes it clearer where an exception has occurred. So this change

- removes the reference to `AIRBRAKE_API_KEY`. It's just not used anywhere anymore
- updates the reference to `AIRBRAKE_FO_PROJECT_KEY`