Update Google Analytics env var
https://eaflood.atlassian.net/browse/RUBY-878

We have found since the move to a new AWS environment in January 2019 Google Analytics data stopped. Investigation found the cause to be that the code is looking for the env var `GOOGLE_TAG` when what the environment contains is `GOOGLE_ANALYTICS_ID`.

Rather than change all our deployments, we've instead opted to change the code.