Add exclusions for state jumper controller
FRAE includes a helper controller that was designed to allow you to quickly jumpt to any page in the journey. Because its nothing to do the production code its never been included in the previous code coverage metrics which might explain why SonarCloud is reporting 6.7% compared to CodeClimates 35%. Simplecov when ran locally is also reporting 35%.

This sodesn't seem to have solved matters (coverage has bumped to 7.7%) but the nature of SonarCloud seems to be I need the branch merged to master. Then I can see full analysis including all files included, which might help us figure out the discrepency.