Switch to SonarCloud from CodeClimate (#303)
* Switch to SonarCloud from CodeClimate

https://sonarcloud.io/organizations/defra/projects

We currently use [Code Climate](https://codeclimate.com/) for our code quality checks and test coverage analysis. However a number of Defra projects have used an internal SonarQube instance, and some public ones have been using SonarCloud (the cloud SASS version of SonarQube).

It looks like we're about to formally adopt SonarCloud and all projects will coalesce onto. So to keep on top of things and continue to be an example to others, we are working through our repos and switching them to SonarCloud.

** Notes

The figuring out of how to do this (because it wasn't straight forward!) was done in

- https://github.com/DEFRA/waste-exemptions-front-office/pull/317
- https://github.com/DEFRA/waste-exemptions-front-office/pull/318

*** Add exclusions for state jumper controller

FRAE includes a helper controller that was designed to allow you to quickly jumpt to any page in the journey. Because its nothing to do the production code its never been included in the previous code coverage metrics which might explain why SonarCloud was reporting 6.7% compared to CodeClimates 35% until these exclusions were added.